+++ tui/model.go
@@ -3,17 +3,10 @@ package tui
 import (
 	"fmt"
 	"os"
-	"os/exec"
-	"path/filepath"
-	"sort"
-	"strconv"
 	"strings"
 	"time"
 
-	"github.com/charmbracelet/bubbles/viewport"
 	tea "github.com/charmbracelet/bubbletea"
-	"github.com/charmbracelet/lipgloss"
-	"github.com/muesli/termenv"
 )
 
 // HistoryEntry records a completed job for the history pane.
@@ -89,64 +82,24 @@ type ProjectInfo struct {
 
 // Model is the bubbletea TUI model for Fabrik.
 type Model struct {
-	// project info for the footer
-	projectInfo ProjectInfo
-	// poll timer
-	pollInterval time.Duration
-	nextPollAt   time.Time
-	pollCount    int
-
-	// active jobs keyed by "owner/repo#N" (or "#N" for single-repo)
-	active map[string]*activeJob
-	// activeNumToKey maps issue number → current job key for LogEvent routing.
-	// When two repos have the same issue number, the last-started job wins.
-	activeNumToKey map[int]string
-
-	// blocked issues keyed by "owner/repo#N" (or "#N" for single-repo).
-	// These are issues held at the dependency gate waiting for blockers to close.
-	blocked map[string]*blockedIssue
-
-	// completed job history (newest last)
-	history []HistoryEntry
-
-	// history viewport for scrolling
-	historyVP viewport.Model
-
-	// terminal dimensions
+	// layout
 	width  int
 	height int
 
-	// spinner frames
-	spinnerFrames []string
-	spinnerIdx    int
-
-	// now (updated on each TickEvent)
-	now time.Time
-
-	// graphqlStats holds the most recent GraphQL rate limit data from the engine.
-	// Zero value (Limit==0) means no data has been received yet.
-	graphqlStats RateLimitStats
-
-	// pluginDir is the Fabrik plugin directory, passed to claude --plugin-dir
-	pluginDir string
-
-	// statusLine shows the latest poll-level log message in the header
-	statusLine string
-
-	// statusMsg is a transient user-facing message shown in the header instead
-	// of statusLine. It is cleared on each TickEvent so it disappears after ~1s.
-	statusMsg string
-
-	// detailPanel controls whether the inline detail panel is shown between the
-	// active and history sections in View().
+	// focus and confirmation state
+	focusPane   pane
+	confirmQuit bool
 	detailPanel bool
 
-	// selection state
-	focusPane    pane
-	activeIdx    int  // index into sorted active issue numbers
-	histIdx      int  // index into history (0 = newest)
-	confirmClear bool // true when waiting for Y/N on clear-all
-	confirmQuit  bool // true when waiting for q/n on quit-with-active-jobs
+	// plugin directory
+	pluginDir string
+
+	// components
+	header  HeaderComponent
+	active  ActivePaneComponent
+	history HistoryPaneComponent
+	detail  DetailPanelComponent
+	footer  FooterComponent
 
 	// mouse double-click detection
 	lastClickAt time.Time
@@ -154,63 +107,39 @@ type Model struct {
 	lastClickY  int
 }
 
-var (
-	titleStyle = lipgloss.NewStyle().
-			Bold(true).
-			Foreground(lipgloss.Color("205"))
-
-	borderStyle = lipgloss.NewStyle().
-			Border(lipgloss.RoundedBorder()).
-			Padding(0, 1)
-
-	successStyle  = lipgloss.NewStyle().Foreground(lipgloss.Color("42"))
-	failStyle     = lipgloss.NewStyle().Foreground(lipgloss.Color("196"))
-	dimStyle      = lipgloss.NewStyle().Foreground(lipgloss.Color("243"))
-	activeStyle   = lipgloss.NewStyle().Foreground(lipgloss.Color("214"))
-	selectedStyle = lipgloss.NewStyle().Background(lipgloss.Color("238"))
-)
-
-// supportsOSC8 returns true when the terminal is known to support OSC 8 hyperlinks.
-// Detection uses environment variables because no universal runtime query exists.
-func supportsOSC8() bool {
-	prog := strings.ToLower(os.Getenv("TERM_PROGRAM"))
-	switch prog {
-	case "iterm.app", "ghostty", "wezterm", "kitty":
-		return true
-	}
-	if strings.Contains(strings.ToLower(os.Getenv("TERM")), "kitty") {
-		return true
-	}
-	return false
-}
-
-// applyOSC8 wraps the first occurrence of title in plain with an OSC 8 hyperlink
-// pointing to boardURL. If title or boardURL is empty, or if the terminal does not
-// support OSC 8, plain is returned unchanged.
-func applyOSC8(plain, title, boardURL string) string {
-	if title == "" || boardURL == "" || !supportsOSC8() {
-		return plain
-	}
-	return strings.Replace(plain, title, termenv.Hyperlink(boardURL, title), 1)
-}
-
 // New creates an initial TUI model.
 // pollSeconds is the configured polling interval.
 // info provides project metadata displayed in the footer.
 // pluginDir is the Fabrik plugin directory passed to claude --plugin-dir (may be empty).
 func New(pollSeconds int, info ProjectInfo, pluginDir string) Model {
-	vp := viewport.New(80, 10)
-	return Model{
-		projectInfo:    info,
-		pollInterval:   time.Duration(pollSeconds) * time.Second,
+	interval := time.Duration(pollSeconds) * time.Second
+	now := time.Now()
+	spinnerFrames := []string{"⠋", "⠙", "⠹", "⠸", "⠼", "⠴", "⠦", "⠧", "⠇", "⠏"}
+
+	active := ActivePaneComponent{
 		active:         make(map[string]*activeJob),
 		activeNumToKey: make(map[int]string),
 		blocked:        make(map[string]*blockedIssue),
-		history:        LoadHistory(),
-		historyVP:      vp,
-		spinnerFrames:  []string{"⠋", "⠙", "⠹", "⠸", "⠼", "⠴", "⠦", "⠧", "⠇", "⠏"},
-		now:            time.Now(),
+		focused:        true,
+		spinnerFrames:  spinnerFrames,
+		now:            now,
 		pluginDir:      pluginDir,
+	}
+
+	return Model{
+		focusPane: paneActive,
+		pluginDir: pluginDir,
+		header: HeaderComponent{
+			pollInterval:  interval,
+			now:           now,
+			fabrikVersion: info.FabrikVersion,
+		},
+		active:  active,
+		history: NewHistoryPaneComponent(),
+		footer: FooterComponent{
+			projectInfo: info,
+			now:         now,
+		},
 	}
 }
 
@@ -232,57 +161,25 @@ func (m Model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 	case tea.KeyMsg:
 		switch ev.String() {
 		case "ctrl+c":
-			// Always quit immediately, bypassing all confirmation dialogs.
 			return m, tea.Quit
 
 		case "q":
-			if m.confirmClear {
-				m.confirmClear = false
+			if m.history.ConfirmClear() {
+				m.history.SetConfirmClear(false)
 				return m, nil
 			}
 			if m.confirmQuit {
 				return m, tea.Quit
 			}
-			if len(m.active) > 0 {
+			if m.active.ActiveCount() > 0 {
 				m.confirmQuit = true
 				return m, nil
 			}
 			return m, tea.Quit
 
-		case "c":
-			// Delete selected history entry
-			if m.focusPane == paneHistory && len(m.history) > 0 {
-				realIdx := len(m.history) - 1 - m.histIdx
-				if realIdx >= 0 && realIdx < len(m.history) {
-					m.history = append(m.history[:realIdx], m.history[realIdx+1:]...)
-					SaveHistory(m.history)
-					if m.histIdx >= len(m.history) && m.histIdx > 0 {
-						m.histIdx--
-					}
-					m.updateHistoryViewport(false)
-				}
-			}
-			return m, nil
-
-		case "C":
-			if m.focusPane == paneHistory && len(m.history) > 0 {
-				if m.confirmClear {
-					// Confirmed — clear all
-					m.history = nil
-					m.histIdx = 0
-					m.confirmClear = false
-					SaveHistory(m.history)
-					m.updateHistoryViewport(false)
-				} else {
-					// Ask for confirmation
-					m.confirmClear = true
-				}
-			}
-			return m, nil
-
 		case "n", "N":
-			if m.confirmClear {
-				m.confirmClear = false
+			if m.history.ConfirmClear() {
+				m.history.SetConfirmClear(false)
 				return m, nil
 			}
 			if m.confirmQuit {
@@ -295,9 +192,8 @@ func (m Model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 			}
 
 		case "esc":
-			// Priority: confirmClear → detailPanel → confirmQuit → trigger quit flow.
-			if m.confirmClear {
-				m.confirmClear = false
+			if m.history.ConfirmClear() {
+				m.history.SetConfirmClear(false)
 				return m, nil
 			}
 			if m.detailPanel {
@@ -308,7 +204,7 @@ func (m Model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 				m.confirmQuit = false
 				return m, nil
 			}
-			if len(m.active) > 0 {
+			if m.active.ActiveCount() > 0 {
 				m.confirmQuit = true
 				return m, nil
 			}
@@ -320,215 +216,143 @@ func (m Model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 			} else {
 				m.focusPane = paneActive
 			}
-			m.updateHistoryViewport(false)
-			return m, nil
-
-		case "up", "k":
-			if m.focusPane == paneActive {
-				if m.activeIdx > 0 {
-					m.activeIdx--
-				}
-			} else {
-				if m.histIdx > 0 {
-					m.histIdx--
-					m.updateHistoryViewport(false)
-				}
-			}
-			return m, nil
-
-		case "down", "j":
-			if m.focusPane == paneActive {
-				keys := m.sortedActiveKeys()
-				if m.activeIdx < len(keys)-1 {
-					m.activeIdx++
-				}
-			} else {
-				if m.histIdx < len(m.history)-1 {
-					m.histIdx++
-					m.updateHistoryViewport(false)
-				}
-			}
-			return m, nil
-
-		case "l":
-			// Launch fabrik watch <issueNumber> inline via tea.ExecProcess.
-			// Suspends the TUI until the user exits watch with q.
-			if m.focusPane == paneActive {
-				keys := m.sortedActiveKeys()
-				if m.activeIdx < len(keys) {
-					if job, ok := m.active[keys[m.activeIdx]]; ok {
-						return m, openWatchInlineCmd(job.IssueNumber, job.Repo)
-					}
-				}
-			} else if m.focusPane == paneHistory && len(m.history) > 0 {
-				realIdx := len(m.history) - 1 - m.histIdx
-				if realIdx >= 0 && realIdx < len(m.history) {
-					h := m.history[realIdx]
-					return m, openWatchInlineCmd(h.IssueNumber, h.Repo)
-				}
-			}
+			m.syncFocus()
+			m.updateLayout(false)
 			return m, nil
 
 		case "enter":
-			// Toggle the inline detail panel.
 			m.detailPanel = !m.detailPanel
 			return m, nil
 
 		case "r":
-			// Active pane: show status message — active sessions must not be interrupted.
-			// History pane: resume an interactive Claude session in the issue's worktree,
-			// but only when the issue is not currently active.
 			if m.focusPane == paneActive {
-				keys := m.sortedActiveKeys()
-				if m.activeIdx < len(keys) {
-					m.statusMsg = "stage in progress — use l to watch"
+				if m.active.SelectedJob() != nil {
+					m.header.SetStatusMsg("stage in progress — use l to watch")
 				}
 				return m, nil
-			} else if m.focusPane == paneHistory && len(m.history) > 0 {
-				realIdx := len(m.history) - 1 - m.histIdx
-				if realIdx >= 0 && realIdx < len(m.history) {
-					h := m.history[realIdx]
-					if m.isActiveIssue(h) {
-						m.statusMsg = "stage in progress — use l to watch"
-						return m, nil
-					}
-					// Check worktree exists before launching resume.
-					cwd, _ := os.Getwd()
-					worktreePath := worktreePathForIssue(cwd, h.Repo, h.IssueNumber)
-					if _, err := os.Stat(worktreePath); os.IsNotExist(err) {
-						m.statusMsg = fmt.Sprintf("no worktree for #%d", h.IssueNumber)
-						return m, nil
-					}
-					return m, m.openResumeInlineCmd(h.Repo, h.IssueNumber, h.StageName, h.StageModel, worktreePath)
+			} else if m.focusPane == paneHistory {
+				entry := m.history.SelectedEntry()
+				if entry == nil {
+					return m, nil
 				}
+				if isActiveIssue(m.active.active, *entry) {
+					m.header.SetStatusMsg("stage in progress — use l to watch")
+					return m, nil
+				}
+				cwd, _ := os.Getwd()
+				worktreePath := worktreePathForIssue(cwd, entry.Repo, entry.IssueNumber)
+				if _, err := os.Stat(worktreePath); os.IsNotExist(err) {
+					m.header.SetStatusMsg(fmt.Sprintf("no worktree for #%d", entry.IssueNumber))
+					return m, nil
+				}
+				return m, openResumeInlineCmd(m.pluginDir, entry.Repo, entry.IssueNumber, entry.StageName, entry.StageModel, worktreePath)
 			}
 			return m, nil
 
+		case "c", "C":
+			if m.focusPane == paneHistory {
+				comp, cmd := m.history.Update(msg)
+				m.history = comp.(HistoryPaneComponent)
+				m.updateLayout(false)
+				return m, cmd
+			}
+			return m, nil
+
+		case "up", "k", "down", "j":
+			if m.focusPane == paneActive {
+				comp, cmd := m.active.Update(msg)
+				m.active = comp.(ActivePaneComponent)
+				return m, cmd
+			}
+			comp, cmd := m.history.Update(msg)
+			m.history = comp.(HistoryPaneComponent)
+			m.updateLayout(false)
+			m.history.ScrollToVisible()
+			return m, cmd
+
+		case "l":
+			if m.focusPane == paneActive {
+				comp, cmd := m.active.Update(msg)
+				m.active = comp.(ActivePaneComponent)
+				return m, cmd
+			}
+			if m.focusPane == paneHistory {
+				comp, cmd := m.history.Update(msg)
+				m.history = comp.(HistoryPaneComponent)
+				return m, cmd
+			}
+			return m, nil
 		}
 		// Forward other key events to the history viewport for scrolling
 		var cmd tea.Cmd
-		m.historyVP, cmd = m.historyVP.Update(msg)
+		m.history.historyVP, cmd = m.history.historyVP.Update(msg)
 		return m, cmd
 
 	case tea.WindowSizeMsg:
 		m.width = ev.Width
 		m.height = ev.Height
-		m.updateHistoryViewport(false)
+		m.updateLayout(false)
 		return m, nil
 
 	case watchExitMsg:
-		// fabrik watch subprocess exited — nothing to do (TUI resumes automatically).
 		return m, nil
 
 	case claudeResumeFinishedMsg:
-		// claude --resume subprocess exited — nothing to do (TUI resumes automatically).
 		return m, nil
 
 	case TickEvent:
-		m.now = ev.At
-		m.spinnerIdx = (m.spinnerIdx + 1) % len(m.spinnerFrames)
-		m.statusMsg = "" // clear transient status message each tick
+		// Fan out to all components
+		comp, _ := m.header.Update(msg)
+		m.header = comp.(HeaderComponent)
+		comp, _ = m.active.Update(msg)
+		m.active = comp.(ActivePaneComponent)
+		comp, _ = m.footer.Update(msg)
+		m.footer = comp.(FooterComponent)
 		return m, tickCmd()
 
 	case ProjectMetaEvent:
-		m.projectInfo.BoardTitle = ev.BoardTitle
-		m.projectInfo.BoardURL = ev.BoardURL
+		comp, _ := m.footer.Update(msg)
+		m.footer = comp.(FooterComponent)
 		return m, nil
 
 	case PollStartedEvent:
-		m.nextPollAt = time.Now().Add(m.pollInterval)
+		comp, _ := m.header.Update(msg)
+		m.header = comp.(HeaderComponent)
 		return m, nil
 
 	case PollCompletedEvent:
-		m.pollCount++
-		m.nextPollAt = time.Now().Add(m.pollInterval)
-		if ev.GraphQLStats.Limit > 0 {
-			m.graphqlStats = ev.GraphQLStats
-		}
+		comp, _ := m.header.Update(msg)
+		m.header = comp.(HeaderComponent)
+		comp, _ = m.footer.Update(msg)
+		m.footer = comp.(FooterComponent)
 		return m, nil
 
 	case JobStartedEvent:
-		key := activeJobKey(ev.Repo, ev.IssueNumber)
-		m.active[key] = &activeJob{
-			IssueNumber: ev.IssueNumber,
-			Repo:        ev.Repo,
-			Title:       ev.Title,
-			StageName:   ev.StageName,
-			IsComment:   ev.IsComment,
-			StartedAt:   ev.StartedAt,
-		}
-		m.activeNumToKey[ev.IssueNumber] = key
-		// Passively remove from blocked map if the issue is now running.
-		delete(m.blocked, key)
+		comp, _ := m.active.Update(msg)
+		m.active = comp.(ActivePaneComponent)
 		return m, nil
 
 	case IssueBlockedEvent:
-		key := activeJobKey(ev.Repo, ev.IssueNumber)
-		m.blocked[key] = &blockedIssue{
-			IssueNumber: ev.IssueNumber,
-			Repo:        ev.Repo,
-			Title:       ev.Title,
-			StageName:   ev.StageName,
-			WaitingFor:  ev.WaitingFor,
-		}
+		comp, _ := m.active.Update(msg)
+		m.active = comp.(ActivePaneComponent)
 		return m, nil
 
 	case JobCompletedEvent:
-		key := activeJobKey(ev.Repo, ev.IssueNumber)
-		delete(m.active, key)
-		if m.activeNumToKey[ev.IssueNumber] == key {
-			delete(m.activeNumToKey, ev.IssueNumber)
-		}
-		entry := HistoryEntry{
-			IssueNumber:    ev.IssueNumber,
-			Repo:           ev.Repo,
-			Title:          ev.Title,
-			StageName:      ev.StageName,
-			StageModel:     ev.StageModel,
-			IsComment:      ev.IsComment,
-			Success:        ev.Success,
-			Completed:      ev.Completed,
-			BlockedOnInput: ev.BlockedOnInput,
-			Duration:       ev.Duration,
-			CompletedAt:    ev.CompletedAt,
-			TurnsUsed:      ev.TurnsUsed,
-			MaxTurns:       ev.MaxTurns,
-			CostUSD:        ev.CostUSD,
-		}
-		// Remove any existing entry with the same dedup key so repeated dispatches
-		// of the same (issue, stage) replace rather than accumulate.
-		dupKey := historyDedupKey{
-			IssueNumber: entry.IssueNumber,
-			Repo:        entry.Repo,
-			StageName:   entry.StageName,
-			IsComment:   entry.IsComment,
-		}
-		filtered := m.history[:0]
-		for _, h := range m.history {
-			k := historyDedupKey{
-				IssueNumber: h.IssueNumber,
-				Repo:        h.Repo,
-				StageName:   h.StageName,
-				IsComment:   h.IsComment,
-			}
-			if k != dupKey {
-				filtered = append(filtered, h)
-			}
-		}
-		m.history = append(filtered, entry)
-		SaveHistory(m.history)
-		m.updateHistoryViewport(true)
+		// Fan out to both active (remove job) and history (add entry)
+		comp, _ := m.active.Update(msg)
+		m.active = comp.(ActivePaneComponent)
+		hcomp, _ := m.history.Update(msg)
+		m.history = hcomp.(HistoryPaneComponent)
+		m.updateLayout(true)
 		return m, nil
 
 	case LogEvent:
 		if ev.IssueNumber == 0 {
-			// Poll-level message — show in header status line
-			m.statusLine = fmt.Sprintf("[%s] %s", ev.Tag, strings.TrimRight(ev.Message, "\n"))
-		} else if key, known := m.activeNumToKey[ev.IssueNumber]; known {
-			if job, ok := m.active[key]; ok {
-				job.LastTag = ev.Tag
-				job.LastLine = strings.TrimRight(ev.Message, "\n")
-			}
+			comp, _ := m.header.Update(msg)
+			m.header = comp.(HeaderComponent)
+		} else {
+			comp, _ := m.active.Update(msg)
+			m.active = comp.(ActivePaneComponent)
 		}
 		return m, nil
 
@@ -539,30 +363,30 @@ func (m Model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 	return m, nil
 }
 
-// handleMouse processes a tea.MouseMsg: forwards all mouse events to the history
-// viewport (which handles wheel scroll internally), then performs hit-testing for
-// left-click events to update selection state and detect double-clicks.
+// handleMouse processes mouse events: forwards wheel scroll to the history
+// viewport, then performs hit-testing using component Height() returns.
 func (m Model) handleMouse(ev tea.MouseMsg) (tea.Model, tea.Cmd) {
 	// Forward all mouse events to the history viewport so wheel-scroll works.
-	var cmd tea.Cmd
-	m.historyVP, cmd = m.historyVP.Update(ev)
+	cmd := m.history.ForwardMouseEvent(ev)
 
 	if ev.Button != tea.MouseButtonLeft || ev.Action != tea.MouseActionPress {
 		return m, cmd
 	}
 
-	// Compute layout Y positions for hit-testing.
-	nActive := len(m.active)
-	activeH := activeHeight(nActive)
+	// Compute layout Y positions from component Height() returns.
+	headerH := m.header.Height()
+	activeH := m.active.Height()
 
-	// histTopY: Y where the history pane border starts.
-	// Active pane occupies Y=1..activeH (inclusive). Detail panel may follow.
-	histTopY := 1 + activeH
+	detailH := 0
 	if m.detailPanel {
-		if detail := m.viewDetail(); detail != "" {
-			histTopY += strings.Count(detail, "\n") + 1
-		}
+		m.prepareDetailItem()
+		detailH = m.detail.Height()
 	}
+
+	activeStart := headerH
+	activeEnd := activeStart + activeH
+	detailEnd := activeEnd + detailH
+	histStart := detailEnd
 
 	clickY := ev.Y
 
@@ -574,68 +398,34 @@ func (m Model) handleMouse(ev tea.MouseMsg) (tea.Model, tea.Cmd) {
 	m.lastClickY = ev.Y
 
 	switch {
-	case clickY >= 1 && clickY <= activeH:
-		// Click anywhere in the active pane (borders, title, or content rows).
+	case clickY >= activeStart && clickY < activeEnd:
 		m.focusPane = paneActive
+		m.syncFocus()
+		localY := clickY - activeStart
+		m.active.HandleClick(ev.X, localY)
 
-		// Content rows start at Y=3 (border-top at Y=1, title at Y=2).
-		// Border-bottom is at Y=activeH; don't treat it as a row selection.
-		if clickY >= 3 && clickY < activeH {
-			keys := m.sortedActiveKeys()
-			nKeys := len(keys)
-			maxLines := activeH - 3
-			start := 0
-			hasEllipsis := false
-			if nKeys > maxLines && maxLines > 0 {
-				start = m.activeIdx - maxLines/2
-				if start < 0 {
-					start = 0
-				}
-				if start+maxLines > nKeys {
-					start = max(nKeys-maxLines, 0)
-				}
-				// Mirror viewActive(): when ellipsis is shown, the last content
-				// row is the ellipsis indicator and must not be treated as a job.
-				if start > 0 || start+maxLines < nKeys {
-					maxLines--
-					hasEllipsis = true
-				}
-			}
-			visibleRow := clickY - 3
-			if hasEllipsis && visibleRow == maxLines {
-				// Clicked the "… N more" ellipsis row; focus the pane but don't select a job.
-				break
-			}
-			actualIdx := start + visibleRow
-			if actualIdx >= 0 && actualIdx < nKeys {
-				m.activeIdx = actualIdx
-				if isDoubleClick {
-					if job, ok := m.active[keys[actualIdx]]; ok {
-						return m, openWatchInlineCmd(job.IssueNumber, job.Repo)
-					}
-				}
+		// Double-click on active content row opens watch
+		if isDoubleClick && localY >= 2 && localY < activeH-1 {
+			if job := m.active.SelectedJob(); job != nil {
+				return m, openWatchInlineCmd(job.IssueNumber, job.Repo)
 			}
 		}
 
-	case clickY >= histTopY && clickY <= histTopY+1:
-		// Click on history pane border-top or title row → switch focus.
+	case clickY >= histStart && clickY <= histStart+1:
 		m.focusPane = paneHistory
-		m.updateHistoryViewport(false)
+		m.syncFocus()
+		m.updateLayout(false)
 
-	case clickY >= histTopY+2:
-		// Click on history viewport content.
-		visibleRow := clickY - (histTopY + 2)
-		newHistIdx := visibleRow + m.historyVP.YOffset
-		if newHistIdx >= 0 && newHistIdx < len(m.history) {
-			m.focusPane = paneHistory
-			m.histIdx = newHistIdx
-			m.updateHistoryViewport(false)
-			if isDoubleClick {
-				realIdx := len(m.history) - 1 - m.histIdx
-				if realIdx >= 0 && realIdx < len(m.history) {
-					h := m.history[realIdx]
-					return m, openWatchInlineCmd(h.IssueNumber, h.Repo)
-				}
+	case clickY >= histStart+2:
+		m.focusPane = paneHistory
+		m.syncFocus()
+		localY := clickY - histStart
+		m.history.HandleClick(ev.X, localY)
+		m.updateLayout(false)
+
+		if isDoubleClick {
+			if entry := m.history.SelectedEntry(); entry != nil {
+				return m, openWatchInlineCmd(entry.IssueNumber, entry.Repo)
 			}
 		}
 	}
@@ -643,110 +433,62 @@ func (m Model) handleMouse(ev tea.MouseMsg) (tea.Model, tea.Cmd) {
 	return m, cmd
 }
 
-// updateHistoryViewport rebuilds the viewport content from the history slice.
-// Uses a pointer receiver so it can be called on the addressable local copy
-// in Update(); the mutations are included in the returned model value.
-// If scrollToTop is true, the viewport is scrolled to the top (newest entry);
-// otherwise the viewport is adjusted so histIdx remains visible.
-func (m *Model) updateHistoryViewport(scrollToTop bool) {
-	innerWidth := max(m.width-6, 20) // account for border + padding
+// syncFocus updates focused state on pane components to match m.focusPane.
+func (m *Model) syncFocus() {
+	m.active.SetFocused(m.focusPane == paneActive)
+	m.history.SetFocused(m.focusPane == paneHistory)
+}
 
-	var lines []string
-	// Show newest entries at the top.
-	for i := len(m.history) - 1; i >= 0; i-- {
-		h := m.history[i]
-		var status, result string
-		if !h.Success {
-			status = failStyle.Render("✗")
-			result = dimStyle.Render("  (error)")
-		} else if h.BlockedOnInput {
-			status = activeStyle.Render("?")
-			result = dimStyle.Render("  (input needed)")
-		} else if !h.Completed {
-			status = dimStyle.Render("↻")
-			result = dimStyle.Render("  (retry)")
-		} else {
-			status = successStyle.Render("✓")
-		}
-		ts := dimStyle.Render(h.CompletedAt.Format("2006-01-02 15:04"))
-		dur := fmtDuration(h.Duration)
-		stats := ""
-		if h.TurnsUsed > 0 || h.CostUSD > 0 {
-			parts := []string{}
-			if h.MaxTurns > 0 {
-				parts = append(parts, fmt.Sprintf("%d/%d turns", h.TurnsUsed, h.MaxTurns))
-			} else if h.TurnsUsed > 0 {
-				parts = append(parts, fmt.Sprintf("%d turns", h.TurnsUsed))
-			}
-			if h.CostUSD > 0 {
-				parts = append(parts, fmt.Sprintf("$%.2f", h.CostUSD))
-			}
-			stats = dimStyle.Render("  " + strings.Join(parts, " "))
-		}
-		titleStr := ""
-		if h.Title != "" {
-			maxTitle := max(innerWidth-60, 10)
-			t := h.Title
-			if runes := []rune(t); len(runes) > maxTitle {
-				t = string(runes[:maxTitle]) + "…"
-			}
-			titleStr = "  " + dimStyle.Render(t)
-		}
-		stageDisplay := h.StageName
-		if h.IsComment {
-			stageDisplay += " 💬"
-		}
-		// Pad stage column manually — emoji width differs from byte count
-		stagePad := 12 - lipgloss.Width(stageDisplay)
-		if stagePad > 0 {
-			stageDisplay += strings.Repeat(" ", stagePad)
-		}
-		line := fmt.Sprintf("#%-5d %s %s %s  %s%s%s%s",
-			h.IssueNumber, stageDisplay, status, dur, ts, stats, result, titleStr)
-		displayIdx := len(m.history) - 1 - i // 0-based index matching histIdx
-		if m.focusPane == paneHistory && displayIdx == m.histIdx {
-			line = selectedStyle.Render(line)
-		}
-		lines = append(lines, line)
+// updateLayout recomputes the history viewport dimensions and content.
+func (m *Model) updateLayout(scrollToTop bool) {
+	activeH := m.active.Height()
+	detailH := 0
+	if m.detailPanel {
+		m.prepareDetailItem()
+		detailH = m.detail.Height()
 	}
-	m.historyVP.SetContent(strings.Join(lines, "\n"))
-
-	// Update viewport height within the overall layout.
-	// Minimum 1 (not 3) so the history pane can shrink on small terminals
-	// without pushing header/footer off screen. viewport.Model panics on
-	// non-positive Height, so the floor of 1 is both a correctness and
-	// safety requirement.
-	//
-	// Compute the overhead dynamically: border-top (1) + title+hint row(s) + border-bottom (1).
-	// The hint is truncated to fit innerWidth, so titleAndHintLines is almost always 1.
-	// The width check below is defense-in-depth in case an edge case slips through.
-	// focusIndicator matches the logic in viewHistory() — keep in sync.
-	focusIndicator := " "
-	if m.focusPane == paneHistory {
-		focusIndicator = "▸"
-	}
-	vpTitle := dimStyle.Render(fmt.Sprintf("%s History (%d)", focusIndicator, len(m.history)))
-	vpHint := m.historyHint(lipgloss.Width(vpTitle), innerWidth)
-	titleAndHintLines := 1
-	if lipgloss.Width(vpTitle+vpHint) > innerWidth {
-		titleAndHintLines = 2
-	}
-	historyHeight := max(m.height-headerHeight()-activeHeight(len(m.active)+len(m.blocked))-footerHeight()-2-titleAndHintLines, 1)
-	m.historyVP.Width = innerWidth
-	m.historyVP.Height = historyHeight
-
+	availableHeight := m.height - m.header.Height() - activeH - detailH - m.footer.Height()
+	m.history.SetLayout(m.width, availableHeight, m.confirmQuit, m.active.ActiveCount())
 	if scrollToTop {
-		// New entry arrived — scroll to top so newest is visible.
-		m.historyVP.GotoTop()
+		m.history.ScrollToTop()
 	} else {
-		// Keep histIdx visible: clamp YOffset so histIdx falls within
-		// [YOffset, YOffset+Height-1]. SetYOffset clamps internally.
-		if m.histIdx < m.historyVP.YOffset {
-			m.historyVP.SetYOffset(m.histIdx)
-		} else if m.histIdx > m.historyVP.YOffset+m.historyVP.Height-1 {
-			m.historyVP.SetYOffset(m.histIdx - m.historyVP.Height + 1)
-		}
+		m.history.ScrollToVisible()
 	}
+}
+
+// prepareDetailItem constructs the DetailItem from the focused pane's selection.
+func (m *Model) prepareDetailItem() {
+	if m.focusPane == paneActive {
+		if job := m.active.SelectedJob(); job != nil {
+			m.detail.SetItem(&DetailItem{
+				IssueNumber: job.IssueNumber,
+				Title:       job.Title,
+				StageName:   job.StageName,
+				IsActive:    true,
+				Elapsed:     m.header.now.Sub(job.StartedAt),
+			})
+		} else {
+			m.detail.SetItem(nil)
+		}
+	} else if entry := m.history.SelectedEntry(); entry != nil {
+		m.detail.SetItem(&DetailItem{
+			IssueNumber:    entry.IssueNumber,
+			Title:          entry.Title,
+			StageName:      entry.StageName,
+			StageModel:     entry.StageModel,
+			Success:        entry.Success,
+			Completed:      entry.Completed,
+			BlockedOnInput: entry.BlockedOnInput,
+			Duration:       entry.Duration,
+			TurnsUsed:      entry.TurnsUsed,
+			MaxTurns:       entry.MaxTurns,
+			CostUSD:        entry.CostUSD,
+			CompletedAt:    entry.CompletedAt,
+		})
+	} else {
+		m.detail.SetItem(nil)
+	}
+	m.detail.SetVisible(m.detailPanel)
 }
 
 // View renders the full TUI.
@@ -757,549 +499,18 @@ func (m Model) View() string {
 
 	var sections []string
 
-	// ── Header pane ──
-	sections = append(sections, m.viewHeader())
+	sections = append(sections, m.header.View(m.width))
+	sections = append(sections, m.active.View(m.width))
 
-	// ── In-Progress pane ──
-	sections = append(sections, m.viewActive())
-
-	// ── Detail panel (toggled by enter) ──
 	if m.detailPanel {
-		if detail := m.viewDetail(); detail != "" {
+		m.prepareDetailItem()
+		if detail := m.detail.View(m.width); detail != "" {
 			sections = append(sections, detail)
 		}
 	}
 
-	// ── History pane ──
-	sections = append(sections, m.viewHistory())
-
-	// ── Footer ──
-	sections = append(sections, m.viewFooter())
+	sections = append(sections, m.history.View(m.width))
+	sections = append(sections, m.footer.View(m.width))
 
 	return strings.Join(sections, "\n")
-}
-
-func (m Model) viewHeader() string {
-	var timer string
-	remaining := time.Until(m.nextPollAt)
-	if remaining <= 0 {
-		timer = "polling..."
-	} else {
-		timer = fmt.Sprintf("poll in %s", fmtDuration(remaining))
-	}
-
-	title := titleStyle.Render("fabrik")
-	if m.projectInfo.FabrikVersion != "" {
-		title = title + " " + dimStyle.Render(m.projectInfo.FabrikVersion)
-	}
-	timerStr := dimStyle.Render(timer)
-
-	// statusMsg is transient (cleared each tick); it takes priority over statusLine.
-	displayStatus := m.statusMsg
-	if displayStatus == "" {
-		displayStatus = m.statusLine
-	}
-
-	status := ""
-	if displayStatus != "" {
-		status = "  " + dimStyle.Render(displayStatus)
-	}
-
-	// title + status on left, timer on right, single line
-	left := title + status
-	leftWidth := lipgloss.Width(left)
-	timerWidth := lipgloss.Width(timerStr)
-	available := m.width - 4 // leading space + align with bordered pane right edge
-	if leftWidth+timerWidth > available {
-		// Truncate status to fit. The rendered structure is:
-		//   " " + title + "  " + dimStyle(s+"…") + gap + timerStr
-		// where "  " (2 chars) is the prefix and "…" (1 char) is the suffix → 3 chars overhead.
-		maxStatus := max(available-lipgloss.Width(title)-timerWidth-3, 0)
-		if maxStatus > 0 && displayStatus != "" {
-			s := displayStatus
-			// Shrink using display width (lipgloss.Width strips ANSI), matching viewFooter pattern.
-			for lipgloss.Width(s) > maxStatus {
-				runes := []rune(s)
-				if len(runes) == 0 {
-					break
-				}
-				s = string(runes[:len(runes)-1])
-			}
-			s += "…"
-			status = "  " + dimStyle.Render(s)
-		} else {
-			status = ""
-		}
-		left = title + status
-		leftWidth = lipgloss.Width(left)
-	}
-	// Align timer with bordered pane content edge (border + padding = 2 chars on right).
-	gap := max(m.width-4-leftWidth-timerWidth, 0)
-	return " " + left + strings.Repeat(" ", gap) + timerStr
-}
-
-func (m Model) viewActive() string {
-	focusIndicator := " "
-	if m.focusPane == paneActive {
-		focusIndicator = "▸"
-	}
-	title := activeStyle.Render(fmt.Sprintf("%s In Progress (%d)", focusIndicator, len(m.active)+len(m.blocked)))
-
-	maxWidth := max(m.width-6, 20) // account for border + padding
-
-	var lines []string
-	keys := m.sortedActiveKeys()
-
-	spinner := m.spinnerFrames[m.spinnerIdx]
-	for idx, key := range keys {
-		job := m.active[key]
-		elapsed := fmtDuration(m.now.Sub(job.StartedAt))
-		tag := ""
-		if job.LastTag != "" {
-			tag = dimStyle.Render(fmt.Sprintf("[%s]", job.LastTag))
-		}
-		msg := ""
-		if job.LastLine != "" {
-			msg = job.LastLine
-		}
-		titleStr := ""
-		if job.Title != "" {
-			titleStr = dimStyle.Render(job.Title) + " "
-		}
-		stageDisplay := job.StageName
-		if job.IsComment {
-			stageDisplay += " 💬"
-		}
-		stagePad := 12 - lipgloss.Width(stageDisplay)
-		if stagePad > 0 {
-			stageDisplay += strings.Repeat(" ", stagePad)
-		}
-		line := fmt.Sprintf("#%-5d %s %s %s  %s%s %s",
-			job.IssueNumber, stageDisplay, spinner, elapsed, titleStr, tag, msg)
-		// Truncate to terminal width to prevent wrapping
-		if runes := []rune(line); len(runes) > maxWidth {
-			line = string(runes[:maxWidth-1]) + "…"
-		}
-		if m.focusPane == paneActive && idx == m.activeIdx {
-			line = selectedStyle.Render(line)
-		}
-		lines = append(lines, line)
-	}
-
-	// Render blocked issues after active jobs.
-	blockedKeys := make([]string, 0, len(m.blocked))
-	for k := range m.blocked {
-		blockedKeys = append(blockedKeys, k)
-	}
-	sort.Strings(blockedKeys)
-	for _, key := range blockedKeys {
-		b := m.blocked[key]
-		waiting := strings.Join(b.WaitingFor, ", ")
-		titleStr := ""
-		if b.Title != "" {
-			titleStr = dimStyle.Render(b.Title) + " "
-		}
-		stagePad := 12 - lipgloss.Width(b.StageName)
-		stageDisplay := b.StageName
-		if stagePad > 0 {
-			stageDisplay += strings.Repeat(" ", stagePad)
-		}
-		line := fmt.Sprintf("🔒#%-4d %s waiting for: %s  %s",
-			b.IssueNumber, stageDisplay, waiting, titleStr)
-		if runes := []rune(line); len(runes) > maxWidth {
-			line = string(runes[:maxWidth-1]) + "…"
-		}
-		lines = append(lines, line)
-	}
-
-	// Cap visible lines to fit within activeHeight
-	totalLines := len(lines)
-	maxLines := activeHeight(len(m.active)+len(m.blocked)) - 3 // subtract title + border
-	if len(lines) > maxLines && maxLines > 0 {
-		// Show lines around the selected index
-		start := m.activeIdx - maxLines/2
-		if start < 0 {
-			start = 0
-		}
-		if start+maxLines > len(lines) {
-			start = max(len(lines)-maxLines, 0)
-		}
-		if start > 0 || start+maxLines < totalLines {
-			maxLines--
-		}
-		lines = lines[start : start+maxLines]
-		if start > 0 || start+maxLines < totalLines {
-			lines = append(lines, dimStyle.Render(fmt.Sprintf("  … %d more", totalLines-maxLines)))
-		}
-	}
-
-	hint := ""
-	if m.focusPane == paneActive && len(m.active) > 0 {
-		hintPlain := "  [l] watch  [enter] details  [tab] history"
-		hintMax := max(maxWidth-lipgloss.Width(title), 0)
-		hintRunes := []rune(hintPlain)
-		for len(hintRunes) > 0 && lipgloss.Width(dimStyle.Render(string(hintRunes)+"…")) > hintMax {
-			hintRunes = hintRunes[:len(hintRunes)-1]
-		}
-		if len(hintRunes) == len([]rune(hintPlain)) {
-			hint = dimStyle.Render(hintPlain)
-		} else if len(hintRunes) > 0 {
-			hint = dimStyle.Render(string(hintRunes) + "…")
-		}
-	}
-	content := title + hint + "\n" + strings.Join(lines, "\n")
-	return borderStyle.Width(m.width - 4).Render(content)
-}
-
-func (m Model) viewHistory() string {
-	innerWidth := max(m.width-6, 20)
-	focusIndicator := " "
-	if m.focusPane == paneHistory {
-		focusIndicator = "▸"
-	}
-	title := dimStyle.Render(fmt.Sprintf("%s History (%d)", focusIndicator, len(m.history)))
-	hint := m.historyHint(lipgloss.Width(title), innerWidth)
-	content := title + hint + "\n" + m.historyVP.View()
-	return borderStyle.Width(m.width - 4).Render(content)
-}
-
-func (m Model) viewFooter() string {
-	// Assemble left side: CWD [· BoardTitle] [· version]
-	parts := []string{}
-	if m.projectInfo.CWD != "" {
-		parts = append(parts, m.projectInfo.CWD)
-	}
-	if m.projectInfo.BoardTitle != "" {
-		parts = append(parts, m.projectInfo.BoardTitle)
-	}
-	if m.projectInfo.Version != "" {
-		parts = append(parts, m.projectInfo.Version)
-	}
-	leftPlain := strings.Join(parts, " · ")
-
-	// Assemble right side: "<remaining>/<limit>  <countdown>" (empty when no data).
-	var rightStr string
-	if m.graphqlStats.Limit > 0 {
-		countdown := fmtRateLimitCountdown(m.graphqlStats.Reset, m.now)
-		plain := fmt.Sprintf("%d/%d  %s", m.graphqlStats.Remaining, m.graphqlStats.Limit, countdown)
-		pct := float64(m.graphqlStats.Remaining) / float64(m.graphqlStats.Limit)
-		var style lipgloss.Style
-		switch {
-		case pct > 0.5:
-			style = successStyle
-		case pct > 0.2:
-			style = activeStyle
-		default:
-			style = failStyle
-		}
-		rightStr = style.Render(plain)
-	}
-
-	maxWidth := m.width - 2 // leave 1 char margin on each side
-	if maxWidth < 1 {
-		maxWidth = 1
-	}
-
-	if rightStr == "" {
-		// No right side — original single-side logic.
-		footer := dimStyle.Render(applyOSC8(leftPlain, m.projectInfo.BoardTitle, m.projectInfo.BoardURL))
-		if lipgloss.Width(footer) > maxWidth {
-			runes := []rune(leftPlain)
-			for len(runes) > 0 && lipgloss.Width(dimStyle.Render(string(runes)+"…")) > maxWidth {
-				runes = runes[:len(runes)-1]
-			}
-			footer = dimStyle.Render(applyOSC8(string(runes)+"…", m.projectInfo.BoardTitle, m.projectInfo.BoardURL))
-		}
-		return footer
-	}
-
-	// Two-sided layout: left + gap + right.
-	// Right side is always preserved; left is truncated when space is tight.
-	rightWidth := lipgloss.Width(rightStr)
-	leftRendered := dimStyle.Render(applyOSC8(leftPlain, m.projectInfo.BoardTitle, m.projectInfo.BoardURL))
-	gap := maxWidth - lipgloss.Width(leftRendered) - rightWidth
-	if gap < 1 {
-		// Left must be truncated to make room for the right side.
-		// Reserve 1 space minimum between sides.
-		availLeft := maxWidth - rightWidth - 1
-		if availLeft < 0 {
-			availLeft = 0
-		}
-		runes := []rune(leftPlain)
-		if availLeft == 0 {
-			leftRendered = ""
-		} else {
-			for len(runes) > 0 && lipgloss.Width(dimStyle.Render(string(runes)+"…")) > availLeft {
-				runes = runes[:len(runes)-1]
-			}
-			if len(runes) == 0 {
-				leftRendered = ""
-			} else {
-				leftRendered = dimStyle.Render(applyOSC8(string(runes)+"…", m.projectInfo.BoardTitle, m.projectInfo.BoardURL))
-			}
-		}
-		gap = maxWidth - lipgloss.Width(leftRendered) - rightWidth
-		if gap < 0 {
-			gap = 0
-		}
-	}
-	return leftRendered + strings.Repeat(" ", gap) + rightStr
-}
-
-// historyHint returns the rendered hint string for the history panel, truncated
-// to fit within innerWidth-titleDisplayWidth columns. It encapsulates all state
-// branches (confirmClear, confirmQuit, normal, empty) and is used by both
-// viewHistory() and updateHistoryViewport() to keep them in sync.
-func (m Model) historyHint(titleDisplayWidth, innerWidth int) string {
-	maxHintWidth := max(innerWidth-titleDisplayWidth, 0)
-	if maxHintWidth == 0 {
-		return ""
-	}
-
-	var plainText string
-	var style lipgloss.Style
-	if m.confirmClear {
-		plainText = "  Clear all history? [C]onfirm / [n]o"
-		style = failStyle
-	} else if m.confirmQuit {
-		plainText = fmt.Sprintf("  Quit Fabrik? %d jobs still in progress \u2014 they will be interrupted.  [q] Quit anyway   [n/Escape] Cancel", len(m.active))
-		style = failStyle
-	} else if m.focusPane == paneHistory && len(m.history) > 0 {
-		plainText = "  [r]esume  [l] watch  [enter] details  [c]lear  [C]lear all  [tab] in-progress"
-		style = dimStyle
-	} else {
-		return ""
-	}
-
-	// Check if it fits as-is.
-	rendered := style.Render(plainText)
-	if lipgloss.Width(rendered) <= maxHintWidth {
-		return rendered
-	}
-
-	// Shrink rune-by-rune until it fits with ellipsis.
-	runes := []rune(plainText)
-	for len(runes) > 0 && lipgloss.Width(style.Render(string(runes)+"…")) > maxHintWidth {
-		runes = runes[:len(runes)-1]
-	}
-	if len(runes) == 0 {
-		return ""
-	}
-	return style.Render(string(runes) + "…")
-}
-
-// headerHeight returns the approximate line height of the header pane.
-func headerHeight() int {
-	return 1
-}
-
-// footerHeight returns the height of the footer pane (one persistent line).
-func footerHeight() int {
-	return 1
-}
-
-// activeHeight returns the approximate line height of the active pane.
-// Capped to avoid pushing history off screen on small terminals.
-func activeHeight(n int) int {
-	base := max(n+1, 2) + 2
-	if base > 10 {
-		return 11
-	}
-	return base
-}
-
-// sortedActiveKeys returns job keys from the active map in sorted order.
-// Keys have the form "owner/repo#N" or "#N" (single-repo).
-func (m Model) sortedActiveKeys() []string {
-	keys := make([]string, 0, len(m.active))
-	for k := range m.active {
-		keys = append(keys, k)
-	}
-	sort.Strings(keys)
-	return keys
-}
-
-// tuiReadSessionID reads the Claude session ID for a given repo, issue and stage.
-// The path logic mirrors engine.ReadSessionID — keep in sync if either changes.
-func tuiReadSessionID(repo string, issueNumber int, stageName string) string {
-	cwd, _ := os.Getwd()
-	base := filepath.Base(stageName)
-	if base == "" || base == "." || base == "/" || base == string(filepath.Separator) {
-		base = "default"
-	}
-	issuePart := fmt.Sprintf("issue-%d", issueNumber)
-	var sessDir string
-	if repo != "" {
-		repoPart := strings.ReplaceAll(repo, "/", "-")
-		sessDir = filepath.Join(cwd, ".fabrik", "sessions", repoPart, issuePart)
-	} else {
-		sessDir = filepath.Join(cwd, ".fabrik", "sessions", issuePart)
-	}
-	data, err := os.ReadFile(filepath.Join(sessDir, base+".session"))
-	if err != nil {
-		return ""
-	}
-	return strings.TrimSpace(string(data))
-}
-
-// fmtDuration formats a duration as MM:SS.
-func fmtDuration(d time.Duration) string {
-	if d < 0 {
-		d = 0
-	}
-	d = d.Round(time.Second)
-	m := int(d.Minutes())
-	s := int(d.Seconds()) % 60
-	return fmt.Sprintf("%02d:%02d", m, s)
-}
-
-// fmtRateLimitCountdown returns a human-readable string describing how long
-// until the rate limit resets relative to now.
-func fmtRateLimitCountdown(reset time.Time, now time.Time) string {
-	remaining := reset.Sub(now)
-	if remaining <= 0 {
-		return "soon"
-	}
-	secs := int(remaining.Seconds())
-	if secs >= 3600 {
-		return fmt.Sprintf("%dh", secs/3600)
-	}
-	if secs >= 60 {
-		return fmt.Sprintf("%dm", secs/60)
-	}
-	return fmt.Sprintf("%ds", secs)
-}
-
-// openWatchInlineCmd returns a tea.Cmd that suspends the TUI and launches
-// "fabrik watch <issueNumber>" in the current terminal via tea.ExecProcess.
-// The TUI is restored automatically when the user exits watch with q.
-// In multi-repo mode, --owner and --repo flags are passed explicitly so the
-// child process watches the correct repository.
-func openWatchInlineCmd(issueNumber int, repo string) tea.Cmd {
-	fabrikBin, err := os.Executable()
-	if err != nil {
-		fabrikBin = "fabrik"
-	}
-	args := []string{"watch"}
-	if repo != "" {
-		parts := strings.SplitN(repo, "/", 2)
-		if len(parts) == 2 {
-			args = append(args, "--owner", parts[0], "--repo", parts[1])
-		}
-	}
-	args = append(args, strconv.Itoa(issueNumber))
-	cmd := exec.Command(fabrikBin, args...)
-	return tea.ExecProcess(cmd, func(err error) tea.Msg {
-		return watchExitMsg{Err: err}
-	})
-}
-
-// openResumeInlineCmd returns a tea.Cmd that suspends the TUI and launches an
-// interactive Claude session in the issue's worktree via tea.ExecProcess.
-// worktreePath must already be verified to exist by the caller.
-// If a session file exists for the given stage, --resume <id> is passed;
-// otherwise a fresh session starts.
-func (m Model) openResumeInlineCmd(repo string, issueNumber int, stageName, stageModel, worktreePath string) tea.Cmd {
-	args := []string{}
-	sessionID := tuiReadSessionID(repo, issueNumber, stageName)
-	if sessionID != "" {
-		args = append(args, "--resume", sessionID)
-	}
-	if stageModel != "" {
-		args = append(args, "--model", stageModel)
-	}
-	if m.pluginDir != "" {
-		args = append(args, "--plugin-dir", m.pluginDir)
-	}
-	cmd := exec.Command("claude", args...)
-	cmd.Dir = worktreePath
-	return tea.ExecProcess(cmd, func(err error) tea.Msg {
-		return claudeResumeFinishedMsg{Err: err}
-	})
-}
-
-// worktreePathForIssue returns the absolute path to the issue's git worktree.
-// Path: <rootDir>/.fabrik/worktrees/<owner>-<repo>/issue-N
-func worktreePathForIssue(rootDir, repo string, issueNumber int) string {
-	issuePart := fmt.Sprintf("issue-%d", issueNumber)
-	if repo != "" {
-		repoPart := strings.ReplaceAll(repo, "/", "-")
-		return filepath.Join(rootDir, ".fabrik", "worktrees", repoPart, issuePart)
-	}
-	return filepath.Join(rootDir, ".fabrik", "worktrees", issuePart)
-}
-
-// isActiveIssue reports whether the history entry's issue is currently being
-// processed (present in the active jobs map). Works correctly in multi-repo mode.
-func (m Model) isActiveIssue(h HistoryEntry) bool {
-	key := activeJobKey(h.Repo, h.IssueNumber)
-	_, ok := m.active[key]
-	return ok
-}
-
-// viewDetail renders an inline detail panel for the currently selected item.
-// Active pane: shows in-flight job metadata.
-// History pane: shows completed job metadata.
-// Returns an empty string when there is nothing to display.
-func (m Model) viewDetail() string {
-	var lines []string
-
-	if m.focusPane == paneActive {
-		keys := m.sortedActiveKeys()
-		if m.activeIdx < len(keys) {
-			if job, ok := m.active[keys[m.activeIdx]]; ok {
-				elapsed := fmtDuration(m.now.Sub(job.StartedAt))
-				lines = append(lines, fmt.Sprintf("Issue:   #%d", job.IssueNumber))
-				if job.Title != "" {
-					lines = append(lines, fmt.Sprintf("Title:   %s", job.Title))
-				}
-				lines = append(lines, fmt.Sprintf("Stage:   %s", job.StageName))
-				lines = append(lines, fmt.Sprintf("Elapsed: %s", elapsed))
-			}
-		}
-	} else if m.focusPane == paneHistory && len(m.history) > 0 {
-		realIdx := len(m.history) - 1 - m.histIdx
-		if realIdx >= 0 && realIdx < len(m.history) {
-			h := m.history[realIdx]
-			statusStr := "success"
-			if !h.Success {
-				statusStr = "error"
-			} else if h.BlockedOnInput {
-				statusStr = "awaiting input"
-			} else if !h.Completed {
-				statusStr = "incomplete"
-			}
-			lines = append(lines, fmt.Sprintf("Issue:    #%d", h.IssueNumber))
-			if h.Title != "" {
-				lines = append(lines, fmt.Sprintf("Title:    %s", h.Title))
-			}
-			lines = append(lines, fmt.Sprintf("Stage:    %s", h.StageName))
-			if h.StageModel != "" {
-				lines = append(lines, fmt.Sprintf("Model:    %s", h.StageModel))
-			}
-			lines = append(lines, fmt.Sprintf("Status:   %s", statusStr))
-			lines = append(lines, fmt.Sprintf("Duration: %s", fmtDuration(h.Duration)))
-			if h.TurnsUsed > 0 {
-				if h.MaxTurns > 0 {
-					lines = append(lines, fmt.Sprintf("Turns:    %d/%d", h.TurnsUsed, h.MaxTurns))
-				} else {
-					lines = append(lines, fmt.Sprintf("Turns:    %d", h.TurnsUsed))
-				}
-			}
-			if h.CostUSD > 0 {
-				lines = append(lines, fmt.Sprintf("Cost:     $%.4f", h.CostUSD))
-			}
-			if !h.CompletedAt.IsZero() {
-				lines = append(lines, fmt.Sprintf("At:       %s", h.CompletedAt.Format("2006-01-02 15:04:05")))
-			}
-		}
-	}
-
-	if len(lines) == 0 {
-		return ""
-	}
-
-	titleLine := dimStyle.Render("Detail") + "  " + dimStyle.Render("[esc] close")
-	content := titleLine + "\n" + strings.Join(lines, "\n")
-	return borderStyle.Width(m.width - 4).Render(content)
 }
