+++ tui/styles.go
@@ -1,0 +1,19 @@
+package tui
+
+import "github.com/charmbracelet/lipgloss"
+
+var (
+	titleStyle = lipgloss.NewStyle().
+			Bold(true).
+			Foreground(lipgloss.Color("205"))
+
+	borderStyle = lipgloss.NewStyle().
+			Border(lipgloss.RoundedBorder()).
+			Padding(0, 1)
+
+	successStyle  = lipgloss.NewStyle().Foreground(lipgloss.Color("42"))
+	failStyle     = lipgloss.NewStyle().Foreground(lipgloss.Color("196"))
+	dimStyle      = lipgloss.NewStyle().Foreground(lipgloss.Color("243"))
+	activeStyle   = lipgloss.NewStyle().Foreground(lipgloss.Color("214"))
+	selectedStyle = lipgloss.NewStyle().Background(lipgloss.Color("238"))
+)
